Changed CLI to use argparse
