+++ .gitignore
@@ -6,3 +6,5 @@ docs/build
 docs/fj.db
 docs/src
 *.db
+.vagrant
+Vagrantfile
