+++ freejournal_cli.py
@@ -1,6 +1,5 @@
 #!/usr/bin/python
 from frontend.cli import commands
-import sys
 
 if __name__ == '__main__':
-    commands.process_command(sys.argv)
+    commands.process_command()
